requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
numpy==1.21.2
opencv-python==4.5.3.56
tensorflow==2.6.0
Pillow==8.3.2
pydantic==1.8.2
pandas==1.3.3
scikit-learn==0.24.2
albumentations==1.1.0
tqdm==4.62.3
